requirements.txt
pandas
matplotlib
streamlit
python-dotenv
openai
networkx
pyvis
pillow
